Made some fixes from after stream:
* Renamed behavior to match name, no biggie
* Added job checking in behavior to see if it was done, and fix for
  OnDisable so reloads wouldn't give errors
* Added `checked` ... check, around noise size calc to avoid int
  overflow
* Moved disposableItems def so it wouldn't be a race condition for
  clearing
* FIXED the stupid GetIndexFromVert for some reason rounding
  incorrectly, so casting to int properly on position components
* Got rid of JobHandleExtended, I don't think it's neccesary atm
* Simple change to JobQueueBuilder to use handle properly (ie, no more
  JHE that was removed)

+++ Assets/codexhere/MarchingCubes/Utils.cs
@@ -10,7 +10,7 @@ namespace codexhere.MarchingCubes {
         * More Info: https://www.youtube.com/watch?v=0V4Y17yWwB0
         */
         public static int GetIndexFromVert(Vector3 position, Vector2Int size) {
-            return (int)(((position.y + (size.y * position.x)) * size.x) + position.z);
+            return (((int)position.y + (size.y * (int)position.x)) * size.x) + (int)position.z;
         }
 
         /**
